requirements.txt
Flask
Flask-Session>=0.3.2,<0.5
werkzeug>=2
requests>=2,<3
identity>=0.2,<0.3
python-dotenv<0.22
python-dotenv
requests
fuzzywuzzy
python-Levenshtein  # Improves fuzzywuzzy performance
pandas
langdetect
xlrd
PyMuPDF
watchdog==2.1.9
openpyxl
lxml
gunicorn==23.0.0